bump kotlin compatibility version to 2.2 commit_hash:eca4e8fd9584da04cc6af9972a22544d8289e425

+++ client/android/div/src/test/java/com/yandex/div/core/expression/ExpressionResolverImplTest.kt
@@ -325,7 +325,7 @@ class ExpressionResolverImplTest {
 
         assertEquals(2000, expression.evaluate(underTest))
 
-        val expressionWithConverter = mutableExpressionWithConverter<Int, String>(
+        val expressionWithConverter = mutableExpressionWithConverter<Long, String>(
             rawExpression = "@{timer}",
             typeHelper = TYPE_HELPER_STRING,
             converter = { it.toString() }
